Fix excercise 1 of labwork 3 displaying the entered number as float

+++ labwork_3/excercise_1/main.c
@@ -35,7 +35,7 @@ int main(void)
     float input = 0.0f;
     scanf("%f", &input);
 
-    printf("Вы ввели значение: %f\n", input);
+    printf("Вы ввели значение: %i\n", (int)input);
 
     int checkedInput = 0;
